refactor: 提高 WebChat 复用性
