--- dashboard/src/components/shared/UiParentCard.vue
@@ -1,20 +1,0 @@
-<script setup lang="ts">
-const props = defineProps({
-  title: String
-});
-</script>
-
-<template>
-  <v-card variant="outlined" elevation="0" class="withbg">
-    <v-card-item>
-      <div class="d-sm-flex align-center justify-space-between">
-        <v-card-title>{{ props.title }}</v-card-title>
-        <slot name="action"></slot>
-      </div>
-    </v-card-item>
-    <v-divider></v-divider>
-    <v-card-text>
-      <slot />
-    </v-card-text>
-  </v-card>
-</template>
