+++ dashboard/src/views/ChatBoxPage.vue
@@ -1,5 +1,5 @@
 <script setup>
-import ChatPage from './ChatPage.vue';
+import Chat from '@/components/chat/Chat.vue'
 import { useCustomizerStore } from '@/stores/customizer';
 const customizer = useCustomizerStore();
 </script>
@@ -9,7 +9,7 @@ const customizer = useCustomizerStore();
         <div
             style="height: 100%; width: 100%; display: flex; flex-direction: column; align-items: center; justify-content: center;">
             <div id="container">
-                <ChatPage :chatbox-mode="true"></ChatPage>
+                <Chat :chatbox-mode="true"></Chat>
             </div>
         </div>
     </v-app>
@@ -18,24 +18,6 @@ const customizer = useCustomizerStore();
 <style scoped>
 #container {
     width: 100%;
-    height: 100%;
-}
-
-@media (min-width: 768px) {
-    #container {
-        min-width: 600px;
-        min-height: 370px;
-        max-width: 1100px;
-        max-height: 860px;
-        padding: 36px;
-    }
-}
-
-@media (max-width: 767px) {
-    #container {
-        width: 100%;
-        height: 100%;
-        padding: 0;
-    }
+    height: 100vh;
 }
 </style>
